requirements.txt
streamlit
python-dotenv
PyPDF2
google-generativeai
reportlab

